requirements.txt
# MIA_IA_SYSTEM - Requirements Complet
# Version: Production Ready - CORRIGÉ
# Python: 3.9+
#
# CORRECTIONS APPLIQUÉES:
# - Ajout asyncio-throttle>=1.0.0 (utilisé dans market_data_feed.py)
# - Versions explicites pour toutes les dépendances critiques
# - Organisation améliorée par catégories
# - Ajout scipy pour calculs statistiques dans performance_tracker

# === CORE DEPENDENCIES ===

# Data Processing
pandas==2.1.4              # Version fixée pour stabilité
numpy==1.26.4              # Compatible avec pandas
pyarrow>=12.0.0           # Parquet support
fastparquet>=0.8.0        # Alternative parquet

# Trading & Market Data
ib_insync==0.9.86         # Version testée et validée
yfinance>=0.2.28          # Backup data source
requests>=2.31.0          # HTTP requests

# Machine Learning
scikit-learn==1.3.0       # Version stable
xgboost>=2.0.0           # Advanced ML (optional)
joblib>=1.3.0            # Model persistence
scipy>=1.11.0            # Scientific computing (pour stats)

# Technical Analysis
ta-lib>=0.6.4           # Optional - indicateurs techniques

# === MONITORING & SYSTEM ===

# System Monitoring
psutil>=5.9.0            # System metrics
watchdog>=3.0.0          # File monitoring

# Scheduling & Async
schedule>=1.2.0          # Task scheduling
asyncio-throttle>=1.0.0  # Rate limiting (AJOUTÉ - utilisé dans market_data_feed)
aiofiles>=23.0.0         # Async file operations

# === LOGGING & NOTIFICATIONS ===

# Enhanced Logging
colorlog==6.7.0          # Colored logs (version fixée)
loguru>=0.7.0           # Alternative logger (optional)

# Notifications
discord.py==2.3.2       # Discord notifications (version fixée)
python-telegram-bot>=20.0 # Telegram (optional alternative)

# === DATA STORAGE ===

# Database
sqlalchemy>=2.0.0       # ORM
# sqlite3 included in Python

# Cache
diskcache>=5.6.0        # Disk-based cache

# === UTILITIES ===

# Configuration
python-dotenv>=1.0.0    # Environment variables
pydantic>=2.0.0        # Data validation (optional)
pyyaml>=6.0.1          # YAML config files

# Time & Date
pytz>=2023.3           # Timezone handling
python-dateutil>=2.8.0 # Date parsing

# === DEVELOPMENT & TESTING ===

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0  # Async tests
pytest-cov>=4.1.0       # Coverage
pytest-mock>=3.11.0     # Mocking

# Code Quality
black>=23.0.0          # Code formatting
flake8>=6.0.0         # Linting
mypy>=1.5.0           # Type checking
isort>=5.12.0         # Import sorting

# Documentation
sphinx>=7.0.0         # Documentation generation
sphinx-rtd-theme>=1.3.0 # ReadTheDocs theme

# === PERFORMANCE ===

# Optimization
numba>=0.58.0         # JIT compilation (optional)
cython>=3.0.0        # C extensions (optional)

# Profiling
memory-profiler>=0.61.0  # Memory profiling
line-profiler>=4.1.0     # Line-by-line profiling

# === DEPLOYMENT ===

# Process Management
supervisor>=4.2.0        # Process supervision (optional)

# Monitoring Production
prometheus-client>=0.17.0 # Metrics export (optional)

# === PLATFORM SPECIFIC ===

# Windows - Sierra Chart Integration
pywin32>=306 ; platform_system == "Windows"  # Windows COM interface
comtypes>=1.2.0 ; platform_system == "Windows"  # COM alternative

# === INSTALLATION NOTES ===

# 1. Installation complète:
#    pip install -r requirements.txt
#
# 2. Installation minimale (production):
#    pip install pandas==2.1.4 numpy==1.26.4 ib_insync==0.9.86 scikit-learn==1.3.0 psutil>=5.9.0
#
# 3. Installation développement:
#    pip install -r requirements.txt
#    pip install -r requirements-dev.txt  # Si existe
#
# 4. Pour TA-Lib (installation séparée requise):
#    - Windows: Download wheel from https://www.lfd.uci.edu/~gohlke/pythonlibs/
#    - Linux: sudo apt-get install ta-lib
#    - Mac: brew install ta-lib
#    Puis: pip install ta-lib
#
# 5. Environnement virtuel recommandé:
#    python -m venv venv
#    source venv/bin/activate  # Linux/Mac
#    venv\Scripts\activate     # Windows
#    pip install -r requirements.txt

# === COMPATIBILITY ===

# Python 3.9+ required
# Tested on:
# - Windows 10/11 (x64)
# - Ubuntu 20.04+ (x64)
# - macOS 12+ (Apple Silicon & Intel)
#
# RAM minimum: 8GB
# RAM recommended: 16GB+